requirements.txt
aiogram>=3.5.0
python-dotenv>=1.0.0
tinydb>=4.8.0
telethon>=1.34.0
